@@ -32,16 +32,14 @@ export 'package:currency_picker/src/currency_utils.dart';
 /// );
 /// ```
 void showCurrencyPicker({
-  @required BuildContext context,
-  @required ValueChanged<Currency> onSelect,
-  List<String> currencyFilter,
-  String searchHint,
+  required BuildContext context,
+  required ValueChanged<Currency> onSelect,
+  List<String>? currencyFilter,
+  String? searchHint,
   bool showFlag = true,
   bool showCurrencyName = true,
   bool showCurrencyCode = true,
 }) {
-  assert(context != null);
-  assert(onSelect != null);
   assert(showCurrencyName || showCurrencyCode,
       'showCurrencyName and showCurrencyCode cannot be both false');
   currency_list.showCurrencyListBottomSheet(
@@ -58,8 +56,8 @@ void showCurrencyPicker({
 @Deprecated('Use showCurrencyPicker instead. '
     'This feature was deprecated after v1.0.2.')
 void showCurrencyListBottomSheet({
-  @required BuildContext context,
-  @required ValueChanged<Currency> onSelect,
+  required BuildContext context,
+  required ValueChanged<Currency> onSelect,
 }) {
   showCurrencyPicker(
     context: context,
