@@ -3,7 +3,7 @@ import 'currency.dart';
 class CurrencyUtils {
   /// Return Flag (Emoji Unicode) of a given currency
   static String countryCodeToEmoji(Currency currency) {
-    String countryCode = currency.flag;
+    final String countryCode = currency.flag;
     // 0x41 is Letter A
     // 0x1F1E6 is Regional Indicator Symbol Letter A
     // Example :
